feat(Contact-Api): add authentication and authorization

+++ src/main/java/com/red/bo/web/contact/ContactController.java
@@ -2,6 +2,7 @@ package com.red.bo.web.contact;
 
 import com.red.bo.core.contact.Contact;
 import com.red.bo.core.contact.ContactService;
+import com.red.bo.core.user.User;
 import com.red.bo.web.contact.mapper.ContactDTO;
 import com.red.bo.web.contact.mapper.ContactMapper;
 import com.red.bo.web.contact.mapper.RequestedContact;
@@ -14,6 +15,7 @@ import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -36,17 +38,18 @@ public class ContactController {
 
     @Operation(description = "Create a new contact")
     @PostMapping
-    public ResponseEntity<ContactDTO> createContact(@Valid @RequestBody RequestedContact contact) {
-        contact.setUser(RequestedUser.builder().id((long)1).build());
+    public ResponseEntity<ContactDTO> createContact(@Valid @RequestBody RequestedContact contact,  @AuthenticationPrincipal User user) {
+        contact.setUser(RequestedUser.builder().id(user.getId()).build());
         Contact createdContact = contactService.createContact(map.toContact(contact));
         return new ResponseEntity<>(map.toContactDTO(createdContact), HttpStatus.CREATED);
     }
 
     @Operation(description = "Get all contacts")
     @GetMapping
-    public ResponseEntity<List<ContactDTO>> getAllContacts() {
+    public ResponseEntity<List<ContactDTO>> getAllContacts(@AuthenticationPrincipal User user) {
         List<ContactDTO> contacts = contactService.getAllContacts()
                 .stream()
+                .filter(contact -> contact.getUser().getId().equals(user.getId()))
                 .map(map::toContactDTO)
                 .toList();
         return new ResponseEntity<>(contacts, HttpStatus.OK);
@@ -56,10 +59,10 @@ public class ContactController {
     @ApiResponse(responseCode = "200", description = "Contact found")
     @ApiResponse(responseCode = "400", description = "Contact not found")
     @GetMapping("/{id}")
-    public ResponseEntity<ContactDTO> getContactById(@NotNull @PathVariable Long id) {
-        ContactDTO contact = map.toContactDTO(contactService.getContactById(id));
-        if (contact != null) {
-            return new ResponseEntity<>(contact, HttpStatus.OK);
+    public ResponseEntity<ContactDTO> getContactById(@NotNull @PathVariable Long id, @AuthenticationPrincipal User user) {
+        Contact contact = contactService.getContactById(id);
+        if (contact != null && contact.getUser().getId().equals(user.getId())) {
+            return new ResponseEntity<>(map.toContactDTO(contact), HttpStatus.OK);
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -67,10 +70,13 @@ public class ContactController {
 
     @Operation(description = "Update a contact by ID")
     @PutMapping("/{id}")
-    public ResponseEntity<ContactDTO> updateContact(@NotNull @PathVariable Long id, @Valid @RequestBody Contact updatedContact) {
-        ContactDTO contact = map.toContactDTO(contactService.updateContact(id, updatedContact));
-        if (contact != null) {
-            return new ResponseEntity<>(contact, HttpStatus.OK);
+    public ResponseEntity<ContactDTO> updateContact(@NotNull @PathVariable Long id,
+                                                    @Valid @RequestBody Contact updatedContact,
+                                                    @AuthenticationPrincipal User user) {
+        Contact contact = contactService.getContactById(id);
+        if (contact != null && contact.getUser().getId().equals(user.getId())) {
+            ContactDTO updatedContactDTO = map.toContactDTO(contactService.updateContact(id, updatedContact));
+            return new ResponseEntity<>(updatedContactDTO, HttpStatus.OK);
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
@@ -78,10 +84,15 @@ public class ContactController {
 
     @Operation(description = "Delete a contact by ID")
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteContact(@NotNull @PathVariable Long id) {
-        boolean deleted = contactService.deleteContact(id);
-        if (deleted) {
-            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+    public ResponseEntity<Void> deleteContact(@NotNull @PathVariable Long id, @AuthenticationPrincipal User user) {
+        Contact contact = contactService.getContactById(id);
+        if (contact != null && contact.getUser().getId().equals(user.getId())) {
+            boolean deleted = contactService.deleteContact(id);
+            if (deleted) {
+                return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
         } else {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
